Use tf-idf instead of raw frequencies
Implement tf-idf with no dampening instead of using raw term frequency
counts. This will compensate for words that occur alot (like "german")
and therefore lacks discriminative power.

Drop the publisherId parameter in the userReadArticle() method. It was
unnecessary to use it since it did not help improve performance.

Add a few simple stop words to the list to improve it.