Fix article updating
The ContestHandler would previously add a new NewsArticle object
regardless of whether the article existed or not. This was not ideal
for updating, so change this mechanism. The recommender now contains
two methods; hasArticle and updateArticle, both of which are used by
the ContestHandler when receiving item update event notifications.